frotz-issue-44: Added compiler support for Apple MacOS Sierra 10.12 plus MacPorts 2.4.1.

+++ src/curses/ux_blorb.h
@@ -21,7 +21,8 @@ typedef struct sampledata_struct {
  */
 typedef struct {
     bb_result_t bbres;
-    ulong type;
+    // ulong type;
+    uint32_t type;
     FILE *fp;
 } myresource;
